Add files via upload
I have created a SuiteLet and User Event that triggers when a Sales Order (SO) is created. Upon creation, an email is sent to the Accounting Manager. The email contains options to either accept or reject the Sales Order. If the Accounting Manager clicks 'Accept,' the status of the Sales Order is changed to 'Accepted.' If 'Reject' is clicked, the status is changed to 'Rejected,' and a reason for the rejection is required. This rejection reason is then populated in the Sales Order.